Move utils to dedicated service
