+++ packages/twenty-server/src/engine/metadata-modules/serverless-function/serverless-function.module.ts
@@ -16,6 +16,7 @@ import { DatabaseEventTriggerEntity } from 'src/engine/metadata-modules/database
 import { WorkspaceManyOrAllFlatEntityMapsCacheModule } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.module';
 import { PermissionsModule } from 'src/engine/metadata-modules/permissions/permissions.module';
 import { RouteTriggerEntity } from 'src/engine/metadata-modules/route-trigger/route-trigger.entity';
+import { ServerlessFunctionBuildModule } from 'src/engine/metadata-modules/serverless-function-build/serverless-function-build.module';
 import { ServerlessFunctionLayerModule } from 'src/engine/metadata-modules/serverless-function-layer/serverless-function-layer.module';
 import { ServerlessFunctionTriggerJob } from 'src/engine/metadata-modules/serverless-function/jobs/serverless-function-trigger.job';
 import { ServerlessFunctionEntity } from 'src/engine/metadata-modules/serverless-function/serverless-function.entity';
@@ -45,6 +46,7 @@ import { WorkspaceMigrationModule } from 'src/engine/workspace-manager/workspace
     PermissionsModule,
     WorkspaceManyOrAllFlatEntityMapsCacheModule,
     WorkspaceMigrationModule,
+    ServerlessFunctionBuildModule,
     ServerlessFunctionLayerModule,
     SubscriptionsModule,
     WorkspaceCacheModule,
