+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration/workspace-migration-runner/action-handlers/workspace-schema-migration-runner-action-handlers.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
 
+import { ServerlessFunctionBuildModule } from 'src/engine/metadata-modules/serverless-function-build/serverless-function-build.module';
 import { WorkspaceSchemaManagerModule } from 'src/engine/twenty-orm/workspace-schema-manager/workspace-schema-manager.module';
 import { CreateAgentActionHandlerService } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-runner/action-handlers/agent/services/create-agent-action-handler.service';
 import { DeleteAgentActionHandlerService } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-runner/action-handlers/agent/services/delete-agent-action-handler.service';
@@ -75,7 +76,7 @@ import { DeleteViewActionHandlerService } from 'src/engine/workspace-manager/wor
 import { UpdateViewActionHandlerService } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-runner/action-handlers/view/services/update-view-action-handler.service';
 
 @Module({
-  imports: [WorkspaceSchemaManagerModule],
+  imports: [WorkspaceSchemaManagerModule, ServerlessFunctionBuildModule],
   providers: [
     CreateFieldActionHandlerService,
     UpdateFieldActionHandlerService,
